requirements.txt
Django==4.2.14
djangorestframework==3.14.0
django-filter==23.5

gunicorn==21.2
